Move command execution to its own interface
